Added nam login and splitter files
